+++ Pinecones/Hydra/src/main/java/com/pinecone/hydra/umc/wolfmc/UlfInstructMessage.java
@@ -6,11 +6,11 @@ import com.pinecone.hydra.umc.msg.InformMessage;
 import com.pinecone.hydra.umc.msg.UMCMethod;
 
 public class UlfInstructMessage extends ArchUMCMessage implements InformMessage {
-    public UlfInstructMessage( UMCMethod method, long controlBits ) {
+    public UlfInstructMessage( UMCMethod method, int controlBits ) {
         super( (Object) null, ExtraEncode.Iussum, method, controlBits );
     }
 
-    public UlfInstructMessage( long controlBits ) {
+    public UlfInstructMessage( int controlBits ) {
         this( UMCMethod.UNDEFINED, controlBits );
     }
 }
